requirements.txt

# Install with: `sudo pip install -r requirements.txt`

openpyxl

xlsxwriter

python-nmap